added location backend integration

+++ app/src/main/java/hackattack/me/hackattack/GameObj.java
@@ -1,14 +1,19 @@
 package hackattack.me.hackattack;
 
+import com.google.firebase.database.Exclude;
+
+import java.util.HashMap;
+import java.util.Map;
+
 /**
  * Created by JakeHoward on 1/21/17.
  */
 
 public class GameObj {
-    public int player1coordx = -1;
-    public int player1coordy = -1;
-    public int player2coordx = -1;
-    public int player2coordy = -1;
+    public double player1coordx = -1;
+    public double player1coordy = -1;
+    public double player2coordx = -1;
+    public double player2coordy = -1;
     public int gameCode = 0;
 
     public GameObj() {
@@ -18,4 +23,16 @@ public class GameObj {
     public GameObj(int gameCode) {
         this.gameCode = gameCode;
     }
+
+//    @Exclude
+//    public Map<String, Object> toMap(){
+//        HashMap<String, Object> result = new HashMap<>();
+//        result.put("player1coordx", player1coordx);
+//        result.put("player1coordy", player1coordy);
+//        result.put("player2coordx", player2coordx);
+//        result.put("player2coordy", player2coordy);
+//        result.put("gameCode", gameCode);
+//
+//        return result;
+//    }
 }
